+++ src/shogun/mathematics/Math.h
@@ -700,8 +700,7 @@ class CMath : public CSGObject
 		static float64_t dot(const float64_t* v1, const float64_t* v2, int32_t n);
 
 		/// compute dot product between v1 and v2 (blas optimized)
-		static float64_t dot(
-			const float32_t* v1, const float32_t* v2, int32_t n);
+		static float32_t dot(const float32_t* v1, const float32_t* v2, int32_t n);
 
 		/// compute dot product between v1 and v2 (for 64bit unsigned ints)
 		static inline float64_t dot(
